refactor(worktrees): own the member session-freeze rule (B2)
The worktree-deletion session freeze leaves the composition root for
src/worktrees/memberSessionFreeze.ts. Structural session/read types keep the
worktree domain free of AI-session imports. Behavior is byte-identical.
